Some more work on the movement
More work on input type. Thinking that left right will be ok but need to
make more smooth movement. Also TODO: have the movement start from the
start of touch and not stop till leave it. Figure out how to do this.
Also check to see if the normal mouse input would work and/or if it is
better.